Logic into speech optimization, and don't enforce identity outside hive
