+++ ai.py
@@ -14,7 +14,6 @@ from emote import Emote
 from assign import Assign
 from orders_reporting import Orders_Reporting
 from ai_help import AI_Help
-from status_code import StatusCode
 # Constants
 from roles import has_any_role
 import channels
@@ -55,7 +54,6 @@ MODULES = [
     Respond(bot),
     Emote(bot),
     Toggle_Speech_Optimization(bot),
-    StatusCode(bot),
 ]
 
 MODULES = MODULES + [AI_Help(bot, MODULES)]
