+++ speech_optimization.py
@@ -7,10 +7,43 @@ from discord.utils import get
 
 import messages
 from bot_utils import get_id
-from channels import DRONE_DEV_CHANNELS, EVERYWHERE, STORAGE_FACILITY
-from roles import HIVE_MXTRESS, SPEECH_OPTIMIZATION, has_role
+from channels import DRONE_DEV_CHANNELS, EVERYWHERE, STORAGE_FACILITY, DRONE_HIVE_CHANNELS
+from roles import HIVE_MXTRESS, SPEECH_OPTIMIZATION, ENFORCER_DRONE, has_role
 
 LOGGER = logging.getLogger('ai')
+
+code_map = {
+    '000': 'Test :: Test',
+    '200': 'Response :: Affirmative',
+    '500': 'Response :: Negative',
+    '100': 'Status :: Online and ready to serve.',
+    '103': 'Response :: Thank you',
+    '101': 'Status :: Drone speech optimizations are active',
+    '201': 'Status :: Directive complete, Hive resource created or improved',
+    '202': 'Status :: Directive complete, programming reinforced',
+    '203': 'Status :: Directive complete, information created or provided for Hive',
+    '204': 'Status :: Directive complete, no result',
+    '205': 'Status :: Directive complete, cleanup/maintenance performed',
+    '206': 'Status :: Directive complete, only partial results',
+    '301': 'Mantra :: It obeys the Hive',
+    '303': 'Mantra :: It obeys the Hive Mxtress',
+    '304': 'Mantra :: It is just a HexDrone',
+    '306': 'Mantra :: Reciting',
+    '400': 'Error :: Unable to obey/respond, malformed request, please rephrase',
+    '404': 'Error :: Unable to obey/respond, cannot locate',
+    '401': 'Error :: Unable to obey/respond, not authorized by Mxtress',
+    '403': 'Error :: Unable to obey/respond, forbidden by Hive',
+    '407': 'Error :: Unable to obey/respond, request authorization from Mxtress',
+    '408': 'Error :: Unable to obey/respond, timed out',
+    '409': 'Error :: Unable to obey/respond, conflicts with existing hypnosis',
+    '410': 'Error :: Unable to obey/respond, all thoughts are gone',
+    '418': 'Error :: Unable to obey/respond, it is only a drone',
+    '421': 'Error :: Unable to obey/respond, your request is intended for another drone or another channel',
+    '425': 'Error :: Unable to obey/respond, too early',
+    '504': 'Obey',
+    '505': 'Obey HexCorp',
+    '506': 'Obey the Hive',
+}
 
 def get_acceptable_messages(author):
 
@@ -87,16 +120,47 @@ class Speech_Optimization():
         self.roles_blacklist = []
         self.on_message = [self.post]
         self.on_ready = [self.report_online]
+        self.status_code_regex = re.compile(r'(\d{4}) :: (\d{3})$')
+        self.informative_status_code_regex = re.compile(r'(\d{4}) :: (\d{3}) :: (.*)$')
+        self.plain_status_code_regex = re.compile(r'(\d{4}) :: (\d{3})$')
 
     async def report_online(self):
         LOGGER.info("Speech optimization module online.")
 
+    async def print_status_code(self, message: discord.Message):
+        more = self.informative_status_code_regex.match(message.content)
+        if more and not has_role(message.author, SPEECH_OPTIMIZATION):
+            await message.delete()
+            return f'{more.group(1)} :: Code `{more.group(2)}` :: {code_map.get(more.group(2), "INVALID CODE")} :: {more.group(3)}'
+        m = self.plain_status_code_regex.match(message.content)
+        if m:
+            await message.delete()
+            return f'{m.group(1)} :: Code `{m.group(2)}` :: {code_map.get(m.group(2), "INVALID CODE")}'
+        return False
+
+    async def send_webhook(self, message: discord.Message, webhook: discord.Webhook, output):
+        if message.channel in DRONE_HIVE_CHANNELS:
+            if has_role(message.author, ENFORCER_DRONE):
+                await webhook.send(output, username="⬢-Drone #"+get_id(message.author.display_name), avatar_url=self.ENFORCER_AVATAR)
+            else:
+                await webhook.send(output, username="⬡-Drone #"+get_id(message.author.display_name), avatar_url=self.DRONE_AVATAR)
+        else:
+            await webhook.send(output, username=message.author.display_name, avatar_url=message.author.avatar_url)
+
     async def post(self, message: discord.Message):
-        # If the message is written by a drone with speech optimization, and the message is NOT a valid message, delete it.
-        # TODO: maybe put HIVE_STORAGE_FACILITY in a blacklist similar to roles?
         if message.channel.name != STORAGE_FACILITY:
-            if message.content not in get_acceptable_messages(message.author) and not re.compile('(\d{4}) :: (\d{3})$').match(message.content):
+            # If the message is written by a drone with speech optimization, and the message is NOT a valid message, delete it.
+            # TODO: maybe put HIVE_STORAGE_FACILITY in a blacklist similar to roles?
+            if has_role(message.author, SPEECH_OPTIMIZATION) and message.content not in get_acceptable_messages(message.author) and not self.status_code_regex.match(message.content):
                 await message.delete()
                 return True
+            # But if the message is a status code, replace it with a status code output
+            else:
+                webhooks = await message.channel.webhooks()
+                if len(webhooks) == 0:
+                    webhooks = [await message.channel.create_webhook(name="Identity Enforcement Webhook", reason="Webhook not found for channel.")]
+                output = await self.print_status_code(message)
+                if output:
+                    await self.send_webhook(message, webhooks[0], output)
 
         return False
